Refactoring code

+++ src/main/java/com/nbp/bear/components/service/NbpUserService.java
@@ -6,9 +6,7 @@ import com.nbp.bear.components.model.NbpUser;
 import com.nbp.bear.components.repository.NbpUserRepository;
 import com.nbp.bear.components.request.NbpUserUpdateRequest;
 import com.nbp.bear.components.response.NbpUserResponse;
-import com.nbp.bear.components.response.NbpUtilResponse;
 import com.nbp.bear.components.util.NbpJwtUtil;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,47 +29,8 @@ public class NbpUserService {
     @Autowired
     private NbpJwtUtil nbpJwtUtil;
 
-    public ResponseEntity<Object> NbpGetPasswordUserService(String email) {
-        try {
-            NbpUser nbpUser = nbpUserRepository.findByEmail(email).get();
-            String nbpPwdGenerator = RandomStringUtils.random(15, NbpConstant.NBP_RANDOM_CHARS);
-            nbpUser.setPassword(bCryptPasswordEncoder.encode(nbpPwdGenerator));
-            nbpUser.setTemporaryPassword(Boolean.FALSE);
-            nbpUserRepository.save(nbpUser);
-            return new ResponseEntity<Object>(new NbpUtilResponse(NbpResponse.NBP_USER_PASSWORD_GENERATED, nbpPwdGenerator), HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_NOT_FOUND, HttpStatus.NOT_FOUND);
-        }
-    }
-
-    public ResponseEntity<Object> NbpResetPasswordUserService(String email,String password) {
-        try {
-            NbpUser nbpUser = nbpUserRepository.findByEmail(email).get();
-            nbpUser.setPassword(bCryptPasswordEncoder.encode(password));
-            nbpUserRepository.save(nbpUser);
-            return new ResponseEntity<Object>(new NbpUtilResponse(NbpResponse.NBP_USER_PASSWORD_GENERATED, nbpUser), HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_NOT_FOUND, HttpStatus.NOT_FOUND);
-        }
-    }
-
-
     public List<NbpUser> NbpGetAllUsers() {
         return nbpUserRepository.findAll();
-    }
-
-    public ResponseEntity<Object> NbpUserRegisterService(NbpUser nbpUser) {
-        Optional<NbpUser> userByUserName = nbpUserRepository.findByUserName(nbpUser.getUserName());
-
-        if (!userByUserName.isPresent()) {
-            Optional<NbpUser> userByEmail = nbpUserRepository.findByEmail(nbpUser.getEmail());
-            if (!userByEmail.isPresent()) {
-                nbpUserRepository.save(nbpUser);
-                return new ResponseEntity<Object>(new NbpUserResponse(NbpResponse.NBP_USER_CREATED, ""), HttpStatus.CREATED);
-            }
-            return new ResponseEntity<Object>(NbpResponse.NBP_USER_EMAIL_EXISTS, HttpStatus.FOUND);
-        }
-        return new ResponseEntity<Object>(NbpResponse.NBP_USER_USERNAME_EXISTS, HttpStatus.FOUND);
     }
 
     public List<String> NbpGetRolesByLoggedInUser(Principal principal) {
@@ -118,9 +77,12 @@ public class NbpUserService {
 
     public ResponseEntity<Object> NbpUserDeleteService(String userId) {
         try {
-            NbpUser nbpUser = nbpUserRepository.findByUserId(userId).get();
-            nbpUserRepository.delete(nbpUser);
-            return new ResponseEntity<Object>(nbpUser, HttpStatus.OK);
+            if(nbpUserRepository.existsById(Integer.parseInt(userId))) {
+                NbpUser nbpUser = nbpUserRepository.findByUserId(userId).get();
+                nbpUserRepository.delete(nbpUser);
+                return new ResponseEntity<Object>(nbpUser, HttpStatus.OK);
+            }
+            return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_NOT_FOUND, HttpStatus.NOT_FOUND);
         } catch (Exception ex) {
             return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_NOT_FOUND, HttpStatus.NOT_FOUND);
         }
@@ -146,22 +108,6 @@ public class NbpUserService {
             nbpUser.setActive(nbpUser.isActive() ? Boolean.FALSE : Boolean.TRUE);
             nbpUserRepository.save(nbpUser);
             return new ResponseEntity<Object>(nbpUser, HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_NOT_FOUND, HttpStatus.NOT_FOUND);
-        }
-    }
-
-    public ResponseEntity<Object> NbpChangePasswordService(int userId, String oldPassword, String newPassword) {
-        try {
-            NbpUser nbpUser = nbpUserRepository.findById(userId).get();
-            if(bCryptPasswordEncoder.matches(oldPassword, nbpUser.getPassword())) {
-                nbpUser.setPassword(bCryptPasswordEncoder.encode(newPassword));
-                nbpUser.setTemporaryPassword(Boolean.TRUE);
-                nbpUserRepository.save(nbpUser);
-                return new ResponseEntity<Object>(new NbpUtilResponse(NbpResponse.NBP_USER_PASSWORD_CHANGED, nbpUser), HttpStatus.OK);
-            } else {
-                return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_PASSWORD_NOT_MATCH, HttpStatus.NOT_FOUND);
-            }
         } catch (Exception ex) {
             return new ResponseEntity<Object>(NbpResponse.NBP_USER_ERROR_NOT_FOUND, HttpStatus.NOT_FOUND);
         }
